+++ src/Akeneo/Channel/Component/Query/PublicApi/Cache/CachedChannelExistsWithLocale.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Akeneo\Channel\Component\Query\PublicApi\Cache;
 
-use Akeneo\Channel\Component\Query\GetChannelCodeWithLocaleCodesInterface;
 use Akeneo\Channel\Component\Query\PublicApi\ChannelExistsWithLocaleInterface;
+use Akeneo\Channel\Component\Query\PublicApi\GetChannelCodeWithLocaleCodesInterface;
 
 /**
  * @author    Nicolas Marniesse <nicolas.marniesse@akeneo.com>